Factored MainMenuAccessPreferences out to itss own class
