+++ collect_app/src/main/java/org/odk/collect/android/preferences/screens/MainMenuAccessPreferencesFragment.java
@@ -1,0 +1,37 @@
+package org.odk.collect.android.preferences.screens;
+
+import android.content.Context;
+import android.os.Bundle;
+
+import androidx.annotation.NonNull;
+
+import org.odk.collect.android.R;
+import org.odk.collect.android.injection.DaggerUtils;
+import org.odk.collect.android.preferences.FormUpdateMode;
+
+import static org.odk.collect.android.configure.SettingsUtils.getFormUpdateMode;
+import static org.odk.collect.android.preferences.keys.AdminKeys.ALLOW_OTHER_WAYS_OF_EDITING_FORM;
+import static org.odk.collect.android.preferences.keys.AdminKeys.KEY_EDIT_SAVED;
+import static org.odk.collect.android.preferences.keys.AdminKeys.KEY_GET_BLANK;
+import static org.odk.collect.android.preferences.utilities.PreferencesUtils.displayDisabled;
+
+public class MainMenuAccessPreferencesFragment extends BaseAdminPreferencesFragment {
+
+    @Override
+    public void onAttach(@NonNull Context context) {
+        super.onAttach(context);
+        DaggerUtils.getComponent(context).inject(this);
+    }
+
+    @Override
+    public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
+        super.onCreatePreferences(savedInstanceState, rootKey);
+        setPreferencesFromResource(R.xml.main_menu_access_preferences, rootKey);
+        findPreference(KEY_EDIT_SAVED).setEnabled(settingsProvider.getAdminSettings().getBoolean(ALLOW_OTHER_WAYS_OF_EDITING_FORM));
+
+        FormUpdateMode formUpdateMode = getFormUpdateMode(requireContext(), settingsProvider.getGeneralSettings());
+        if (formUpdateMode == FormUpdateMode.MATCH_EXACTLY) {
+            displayDisabled(findPreference(KEY_GET_BLANK), false);
+        }
+    }
+}
